Moving adaptive pool op to separate folder in experimental path
