+++ ttnn/cpp/ttnn/operations/pool/generic/generic_pools.cpp
@@ -27,621 +27,6 @@
 namespace ttnn {
 namespace operations::pool {
 
-// Structure for PyTorch's exact adaptive pooling regions
-struct AdaptiveRegion {
-    uint32_t output_h, output_w;              // Output position
-    uint32_t start_h, end_h, start_w, end_w;  // Input region boundaries
-    uint32_t kernel_h, kernel_w;              // Region size
-};
-
-// SAFE HYBRID PADDING+DILATION approach with PyTorch compatibility
-enum class AdaptivePoolStrategy {
-    PURE_UNIFORM,       // Already uniform - no adjustment needed
-    PADDING_DOMINANT,   // Legacy padding approach for compatibility (default for edge cases)
-    DILATION_DOMINANT,  // Conservative dilation for low variance cases
-    COMBINED_OPTIMAL    // Balanced approach with compatibility checks
-};
-
-struct HybridAdaptiveConfig {
-    AdaptivePoolStrategy strategy;
-
-    // Padding configuration
-    std::array<uint32_t, 4> padding;  // [pad_top, pad_bottom, pad_left, pad_right]
-
-    // Dilation configuration
-    std::array<uint32_t, 2> dilation;  // [dilation_h, dilation_w]
-
-    // Resulting uniform parameters
-    std::array<uint32_t, 2> kernel_size;
-    std::array<uint32_t, 2> stride;
-
-    // Analysis metrics
-    uint32_t h_variance, w_variance;              // Kernel variance before adjustment
-    uint32_t h_variance_after, w_variance_after;  // Kernel variance after adjustment
-    double coverage_improvement_percent;          // Improvement over legacy approach (conservative)
-    double memory_overhead_percent;               // Total memory cost
-
-    bool is_beneficial() const {
-        return coverage_improvement_percent > 0.0;  // Any improvement is considered beneficial
-    }
-};
-
-// Generate PyTorch-exact region specifications for true adaptive pooling
-static std::vector<AdaptiveRegion> calculate_pytorch_regions(
-    uint32_t input_h, uint32_t input_w, uint32_t output_h, uint32_t output_w) {
-    std::vector<AdaptiveRegion> regions;
-
-    for (uint32_t out_h = 0; out_h < output_h; out_h++) {
-        // PyTorch's exact height calculation
-        uint32_t start_h = (out_h * input_h) / output_h;
-        uint32_t end_h = ((out_h + 1) * input_h + output_h - 1) / output_h;
-
-        for (uint32_t out_w = 0; out_w < output_w; out_w++) {
-            // PyTorch's exact width calculation
-            uint32_t start_w = (out_w * input_w) / output_w;
-            uint32_t end_w = ((out_w + 1) * input_w + output_w - 1) / output_w;
-
-            regions.push_back(
-                {.output_h = out_h,
-                 .output_w = out_w,
-                 .start_h = start_h,
-                 .end_h = end_h,
-                 .start_w = start_w,
-                 .end_w = end_w,
-                 .kernel_h = end_h - start_h,
-                 .kernel_w = end_w - start_w});
-        }
-    }
-
-    log_info(
-        tt::LogOp,
-        "[PyTorch Exact Regions] Generated {} regions for {}x{} -> {}x{}",
-        regions.size(),
-        input_h,
-        input_w,
-        output_h,
-        output_w);
-
-    return regions;
-}
-
-// Calculate adaptive kernel patterns for analysis
-static std::pair<std::vector<uint32_t>, std::vector<uint32_t>> analyze_adaptive_kernels(
-    uint32_t input_h, uint32_t input_w, uint32_t output_h, uint32_t output_w) {
-    std::vector<uint32_t> h_kernels, w_kernels;
-
-    // Height kernel pattern
-    for (uint32_t out_h = 0; out_h < output_h; out_h++) {
-        uint32_t start_h = (out_h * input_h) / output_h;
-        uint32_t end_h = ((out_h + 1) * input_h + output_h - 1) / output_h;
-        h_kernels.push_back(end_h - start_h);
-    }
-
-    // Width kernel pattern
-    for (uint32_t out_w = 0; out_w < output_w; out_w++) {
-        uint32_t start_w = (out_w * input_w) / output_w;
-        uint32_t end_w = ((out_w + 1) * input_w + output_w - 1) / output_w;
-        w_kernels.push_back(end_w - start_w);
-    }
-
-    return {h_kernels, w_kernels};
-}
-
-// Calculate variance and uniformity metrics
-static std::pair<uint32_t, uint32_t> calculate_kernel_variance(const std::vector<uint32_t>& kernels) {
-    if (kernels.empty()) {
-        return {0, 0};
-    }
-
-    uint32_t min_kernel = *std::min_element(kernels.begin(), kernels.end());
-    uint32_t max_kernel = *std::max_element(kernels.begin(), kernels.end());
-
-    return {min_kernel, max_kernel - min_kernel};  // {min_size, variance}
-}
-
-// PATTERN-BASED HYBRID APPROACH: Analyze kernel patterns and fix with padding/dilation
-static HybridAdaptiveConfig calculate_pattern_based_hybrid_config(
-    uint32_t input_h, uint32_t input_w, uint32_t output_h, uint32_t output_w) {
-    HybridAdaptiveConfig config = {};
-
-    // Step 1: Analyze the current kernel patterns
-    auto [h_kernels, w_kernels] = analyze_adaptive_kernels(input_h, input_w, output_h, output_w);
-    auto [h_min, h_variance] = calculate_kernel_variance(h_kernels);
-    auto [w_min, w_variance] = calculate_kernel_variance(w_kernels);
-
-    config.h_variance = h_variance;
-    config.w_variance = w_variance;
-
-    log_info(
-        tt::LogOp,
-        "[PATTERN HYBRID] Analyzing {}x{} -> {}x{}: h_kernels=[{}...], w_kernels=[{}...], variance={}x{}",
-        input_h,
-        input_w,
-        output_h,
-        output_w,
-        h_kernels.empty() ? 0 : h_kernels[0],
-        w_kernels.empty() ? 0 : w_kernels[0],
-        h_variance,
-        w_variance);
-
-    // Step 2: Calculate basic uniform parameters to test if they work
-    uint32_t base_stride_h = input_h / output_h;
-    uint32_t base_stride_w = input_w / output_w;
-    uint32_t base_kernel_h = (input_h + output_h - 1) / output_h;  // ceil
-    uint32_t base_kernel_w = (input_w + output_w - 1) / output_w;  // ceil
-
-    // Step 3: Check if base uniform approach works correctly
-    uint32_t expected_out_h = (input_h - base_kernel_h) / base_stride_h + 1;
-    uint32_t expected_out_w = (input_w - base_kernel_w) / base_stride_w + 1;
-    bool uniform_works = (expected_out_h == output_h) && (expected_out_w == output_w);
-
-    log_info(
-        tt::LogOp,
-        "[PATTERN HYBRID] Base uniform: kernel={}x{}, stride={}x{}, output={}x{} (expected {}x{}) -> {}",
-        base_kernel_h,
-        base_kernel_w,
-        base_stride_h,
-        base_stride_w,
-        expected_out_h,
-        expected_out_w,
-        output_h,
-        output_w,
-        uniform_works ? "WORKS" : "NEEDS_FIX");
-
-    // Step 4: Determine strategy based on variance and uniform feasibility
-    if (uniform_works && h_variance == 0 && w_variance == 0) {
-        // Perfect case - uniform works and no variance
-        config.strategy = AdaptivePoolStrategy::PURE_UNIFORM;
-        config.kernel_size = {base_kernel_h, base_kernel_w};
-        config.stride = {base_stride_h, base_stride_w};
-        config.padding = {0, 0, 0, 0};
-        config.dilation = {1, 1};
-        config.coverage_improvement_percent = 0.0;
-        config.memory_overhead_percent = 0.0;
-
-        log_info(tt::LogOp, "[PATTERN HYBRID] PURE_UNIFORM selected - base approach works perfectly");
-
-    } else if (h_variance <= 1 && w_variance <= 1) {
-        // Low variance cases - apply pattern-specific optimizations
-        config.strategy = AdaptivePoolStrategy::COMBINED_OPTIMAL;
-
-        // Use the max kernels for uniformity
-        uint32_t target_kernel_h =
-            h_kernels.empty() ? base_kernel_h : *std::max_element(h_kernels.begin(), h_kernels.end());
-        uint32_t target_kernel_w =
-            w_kernels.empty() ? base_kernel_w : *std::max_element(w_kernels.begin(), w_kernels.end());
-
-        config.kernel_size = {target_kernel_h, target_kernel_w};
-
-        // For pattern-detected cases, use optimized stride calculation instead of base
-        uint32_t optimized_stride_h = base_stride_h;
-        uint32_t optimized_stride_w = base_stride_w;
-        config.dilation = {1, 1};
-
-        // Smart padding based on detected kernel patterns
-        uint32_t pad_h = 0, pad_w = 0;
-
-        // Pattern: Edge kernels are smaller (like [13,14,14,14,13])
-        if (w_variance == 1 && w_kernels.size() >= 3) {
-            uint32_t first_w = w_kernels[0];
-            uint32_t last_w = w_kernels[w_kernels.size() - 1];
-            uint32_t middle_w = w_kernels[w_kernels.size() / 2];
-
-            // Classic [small, big, big, big, small] pattern
-            if (first_w == last_w && first_w == middle_w - 1) {
-                pad_w = 2;  // Add 1 pixel padding on each side
-
-                // CRITICAL: Use optimized stride for PyTorch compatibility
-                // The key insight is that stride should be based on the dominant (middle) kernels
-                // For [13,14,14,14,13] pattern, we want stride to work with kernel=14
-                // The working formula: stride_w = middle_kernel - 1 = 14 - 1 = 13
-                optimized_stride_w = middle_w - 1;
-
-                log_info(
-                    tt::LogOp,
-                    "[PATTERN HYBRID] Detected edge-smaller width pattern: [{},{},{},{},...] -> pad_w=2, stride_w={}",
-                    w_kernels.size() > 0 ? w_kernels[0] : 0,
-                    w_kernels.size() > 1 ? w_kernels[1] : 0,
-                    w_kernels.size() > 2 ? w_kernels[2] : 0,
-                    w_kernels.size() > 3 ? w_kernels[3] : 0,
-                    optimized_stride_w);
-            }
-        }
-
-        // Same pattern for height
-        if (h_variance == 1 && h_kernels.size() >= 3) {
-            uint32_t first_h = h_kernels[0];
-            uint32_t last_h = h_kernels[h_kernels.size() - 1];
-            uint32_t middle_h = h_kernels[h_kernels.size() / 2];
-
-            if (first_h == last_h && first_h == middle_h - 1) {
-                pad_h = 2;
-                optimized_stride_h = middle_h - 1;
-                log_info(
-                    tt::LogOp,
-                    "[PATTERN HYBRID] Detected edge-smaller height pattern -> pad_h=2, stride_h={}",
-                    optimized_stride_h);
-            }
-        }
-
-        // Apply the optimized stride
-        config.stride = {optimized_stride_h, optimized_stride_w};
-
-        // Distribute padding symmetrically
-        config.padding = {
-            pad_h / 2,          // top
-            pad_h - pad_h / 2,  // bottom
-            pad_w / 2,          // left
-            pad_w - pad_w / 2   // right
-        };
-
-        config.coverage_improvement_percent = (pad_h > 0 || pad_w > 0) ? 25.0 : 0.0;
-        config.memory_overhead_percent = (pad_h + pad_w) * 100.0 / (input_h + input_w);
-
-        log_info(
-            tt::LogOp,
-            "[PATTERN HYBRID] COMBINED_OPTIMAL selected - kernel={}x{}, stride={}x{} (optimized), "
-            "padding=[{},{},{},{}]",
-            target_kernel_h,
-            target_kernel_w,
-            optimized_stride_h,
-            optimized_stride_w,
-            config.padding[0],
-            config.padding[1],
-            config.padding[2],
-            config.padding[3]);
-
-    } else {
-        // Medium variance - use balanced but conservative approach
-        config.strategy = AdaptivePoolStrategy::COMBINED_OPTIMAL;
-
-        // Use PyTorch-compatible calculations
-        config.stride = {input_h / output_h, input_w / output_w};
-        config.kernel_size = {
-            (input_h + output_h - 1) / output_h,  // ceil division
-            (input_w + output_w - 1) / output_w   // ceil division
-        };
-
-        // Very light padding only if absolutely needed
-        config.padding = {0, 0, 0, 0};
-        config.dilation = {1, 1};
-
-        config.coverage_improvement_percent = 5.0;
-        config.memory_overhead_percent = 0.0;
-
-        log_info(tt::LogOp, "[SAFE HYBRID] COMBINED_OPTIMAL selected (conservative)");
-    }
-
-    // Step 4: Always verify that our parameters produce correct output dimensions
-    uint32_t padded_h = input_h + config.padding[0] + config.padding[1];
-    uint32_t padded_w = input_w + config.padding[2] + config.padding[3];
-    uint32_t final_out_h = (padded_h - config.kernel_size[0]) / config.stride[0] + 1;
-    uint32_t final_out_w = (padded_w - config.kernel_size[1]) / config.stride[1] + 1;
-
-    if (final_out_h != output_h || final_out_w != output_w) {
-        log_warning(
-            tt::LogOp,
-            "[PATTERN HYBRID] Final validation failed! Expected {}x{} but calculated {}x{}, using general legacy "
-            "approach",
-            output_h,
-            output_w,
-            final_out_h,
-            final_out_w);
-
-        // Fallback to simple safe approach when validation fails
-        config.strategy = AdaptivePoolStrategy::PADDING_DOMINANT;
-        config.kernel_size = {base_kernel_h, base_kernel_w};
-        config.stride = {base_stride_h, base_stride_w};
-        config.padding = {0, 0, 0, 0};
-        config.dilation = {1, 1};
-        config.coverage_improvement_percent = 0.0;
-        config.memory_overhead_percent = 0.0;
-    }
-
-    // Step 5: Calculate final variance metrics
-    config.h_variance_after = h_variance;
-    config.w_variance_after = w_variance;
-
-    log_info(
-        tt::LogOp,
-        "[PATTERN HYBRID] FINAL: strategy={}, kernel={}x{}, stride={}x{}, pad=[{},{},{},{}], dilation={}x{}",
-        static_cast<int>(config.strategy),
-        config.kernel_size[0],
-        config.kernel_size[1],
-        config.stride[0],
-        config.stride[1],
-        config.padding[0],
-        config.padding[1],
-        config.padding[2],
-        config.padding[3],
-        config.dilation[0],
-        config.dilation[1]);
-
-    return config;
-}
-
-// Legacy structure for backward compatibility
-struct AdaptivePoolingParams {
-    std::array<uint32_t, 2> kernel_size;
-    std::array<uint32_t, 2> stride;
-    std::array<uint32_t, 4> padding;  // [pad_top, pad_bottom, pad_left, pad_right]
-    uint32_t padded_h;
-    uint32_t padded_w;
-    double memory_overhead_percent;
-};
-
-// Convert hybrid config to legacy format for compatibility
-static AdaptivePoolingParams convert_hybrid_to_legacy(
-    const HybridAdaptiveConfig& hybrid_config, uint32_t input_h, uint32_t input_w) {
-    AdaptivePoolingParams legacy = {};
-
-    legacy.kernel_size = hybrid_config.kernel_size;
-    legacy.stride = hybrid_config.stride;
-    legacy.padding = hybrid_config.padding;
-    legacy.padded_h = input_h + hybrid_config.padding[0] + hybrid_config.padding[1];
-    legacy.padded_w = input_w + hybrid_config.padding[2] + hybrid_config.padding[3];
-    legacy.memory_overhead_percent = hybrid_config.memory_overhead_percent;
-
-    return legacy;
-}
-
-// General algorithm to calculate uniform padding for any adaptive pooling case
-static AdaptivePoolingParams calculate_uniform_adaptive_params_general(
-    uint32_t input_h, uint32_t input_w, uint32_t output_h, uint32_t output_w) {
-    // Calculate PyTorch's exact regions to understand kernel/stride patterns
-    std::vector<uint32_t> h_kernels, w_kernels, h_strides, w_strides;
-
-    // Height analysis
-    for (uint32_t out_h = 0; out_h < output_h; out_h++) {
-        uint32_t start_h = (out_h * input_h) / output_h;
-        uint32_t end_h = ((out_h + 1) * input_h + output_h - 1) / output_h;
-        h_kernels.push_back(end_h - start_h);
-
-        if (out_h > 0) {
-            uint32_t prev_start = ((out_h - 1) * input_h) / output_h;
-            h_strides.push_back(start_h - prev_start);
-        }
-    }
-
-    // Width analysis
-    for (uint32_t out_w = 0; out_w < output_w; out_w++) {
-        uint32_t start_w = (out_w * input_w) / output_w;
-        uint32_t end_w = ((out_w + 1) * input_w + output_w - 1) / output_w;
-        w_kernels.push_back(end_w - start_w);
-
-        if (out_w > 0) {
-            uint32_t prev_start = ((out_w - 1) * input_w) / output_w;
-            w_strides.push_back(start_w - prev_start);
-        }
-    }
-
-    // Find target uniform sizes (use most common or maximum)
-    uint32_t target_kernel_h = *std::max_element(h_kernels.begin(), h_kernels.end());
-    uint32_t target_kernel_w = *std::max_element(w_kernels.begin(), w_kernels.end());
-
-    // For strides, use the most common stride (usually the middle ones)
-    uint32_t target_stride_h = h_strides.empty() ? input_h / output_h : h_strides.at(0);
-    uint32_t target_stride_w = w_strides.empty() ? input_w / output_w : w_strides.at(0);
-
-    // If we have multiple strides, prefer the most common one
-    if (h_strides.size() > 1) {
-        target_stride_h = h_strides.at(h_strides.size() / 2);  // Use middle stride
-    }
-    if (w_strides.size() > 1) {
-        target_stride_w = w_strides.at(w_strides.size() / 2);  // Use middle stride
-    }
-
-    // Calculate padding needed to make edge regions match target
-    uint32_t first_h_diff = (h_kernels.size() > 0) ? target_kernel_h - h_kernels.at(0) : 0;
-    uint32_t last_h_diff = (h_kernels.size() > 0) ? target_kernel_h - h_kernels.at(h_kernels.size() - 1) : 0;
-    uint32_t first_w_diff = (w_kernels.size() > 0) ? target_kernel_w - w_kernels.at(0) : 0;
-    uint32_t last_w_diff = (w_kernels.size() > 0) ? target_kernel_w - w_kernels.at(w_kernels.size() - 1) : 0;
-
-    // Calculate padding needed to make all regions uniform
-    uint32_t pad_h_needed = std::max(first_h_diff, last_h_diff);
-    uint32_t pad_w_needed = std::max(first_w_diff, last_w_diff);
-
-    // Initialize padding distribution
-    uint32_t pad_top = 0, pad_bottom = 0, pad_left = 0, pad_right = 0;
-
-    // Height padding distribution
-    if (pad_h_needed > 0) {
-        pad_top = pad_h_needed / 2;
-        pad_bottom = pad_h_needed - pad_top;
-    }
-
-    // Width padding distribution - handle the specific [13,14,14,14,13] pattern
-    if (pad_w_needed > 0) {
-        // Special pattern: first and last kernels are 1 smaller than target
-        // This matches the user's insight for 64x64->3x5 case
-        if (w_kernels.size() >= 3 && first_w_diff == 1 && last_w_diff == 1) {
-            pad_left = 1;
-            pad_right = 1;
-        } else {
-            // General symmetric padding
-            pad_left = pad_w_needed / 2;
-            pad_right = pad_w_needed - pad_left;
-        }
-    }
-
-    uint32_t actual_padded_h = input_h + pad_top + pad_bottom;
-    uint32_t actual_padded_w = input_w + pad_left + pad_right;
-
-    double memory_overhead =
-        100.0 * (actual_padded_h * actual_padded_w - input_h * input_w) / (double)(input_h * input_w);
-
-    log_info(
-        tt::LogOp,
-        "[General Adaptive] {}x{} -> {}x{}: kernel={}x{}, stride={}x{}, pad=({},{},{},{})",
-        input_h,
-        input_w,
-        output_h,
-        output_w,
-        target_kernel_h,
-        target_kernel_w,
-        target_stride_h,
-        target_stride_w,
-        pad_top,
-        pad_bottom,
-        pad_left,
-        pad_right);
-
-    return AdaptivePoolingParams{
-        .kernel_size = {target_kernel_h, target_kernel_w},
-        .stride = {target_stride_h, target_stride_w},
-        .padding = {pad_top, pad_bottom, pad_left, pad_right},
-        .padded_h = actual_padded_h,
-        .padded_w = actual_padded_w,
-        .memory_overhead_percent = memory_overhead};
-}
-
-// Legacy function - now uses the general algorithm
-static AdaptivePoolingParams calculate_uniform_adaptive_params(
-    uint32_t input_h, uint32_t input_w, uint32_t output_h, uint32_t output_w) {
-    // Calculate base kernel/stride using ceil/floor approach
-    uint32_t kernel_h = (input_h + output_h - 1) / output_h;  // ceil
-    uint32_t stride_h = input_h / output_h;                   // floor
-    uint32_t kernel_w = (input_w + output_w - 1) / output_w;  // ceil
-    uint32_t stride_w = input_w / output_w;                   // floor
-
-    uint32_t pad_top = 0, pad_bottom = 0, pad_left = 0, pad_right = 0;
-
-    // Check if we should apply the optimal rounding pattern
-    // This happens when the basic ceil/floor approach gives poor accuracy due to:
-    // 1. Non-integer input/output ratios (common cause of accuracy issues)
-    // 2. Significant stride mismatch between height and width
-    // 3. Coverage issues (last window extends beyond input)
-
-    // Check for non-integer ratios - a key indicator of accuracy issues
-    // Use remainder to detect non-integer division
-    bool h_non_integer = (input_h % output_h) != 0;
-    bool w_non_integer = (input_w % output_w) != 0;
-
-    // Check for stride mismatch (suggests suboptimal uniform approach)
-    bool stride_mismatch = (stride_h > stride_w + 1) || (stride_w > stride_h + 1);
-
-    // Check for coverage issues
-    uint32_t last_window_start_h = (output_h - 1) * stride_h;
-    uint32_t last_window_end_h = last_window_start_h + kernel_h;
-    uint32_t last_window_start_w = (output_w - 1) * stride_w;
-    uint32_t last_window_end_w = last_window_start_w + kernel_w;
-    bool coverage_issue = (last_window_end_h > input_h) || (last_window_end_w > input_w);
-
-    // Apply optimal pattern if conditions indicate accuracy issues
-    // But avoid applying to cases that are already working well
-    bool needs_improvement = (h_non_integer || w_non_integer || stride_mismatch || coverage_issue);
-
-    // Don't apply improvements to cases with perfect coverage
-    // Perfect coverage means the last window ends exactly at the input boundary
-    bool perfect_coverage = (last_window_end_h == input_h && last_window_end_w == input_w);
-
-    if (needs_improvement && !perfect_coverage) {
-        // ACCURACY FIRST APPROACH: Apply the specific high-accuracy patterns
-        // discovered through systematic testing (matches Python results)
-
-        // For 8×9 → 3×4: use kernel=4×3, stride=3×2, pad=2 (79.3% improvement)
-        if (input_h == 8 && input_w == 9 && output_h == 3 && output_w == 4) {
-            kernel_h = 4;
-            kernel_w = 3;
-            stride_h = 3;
-            stride_w = 2;
-            pad_top = 1;
-            pad_bottom = 1;
-            pad_left = 0;
-            pad_right = 0;
-        }
-        // For 64×64 → 3×5: use optimal asymmetric padding for PCC > 0.985 (99.93% accuracy!)
-        else if (input_h == 64 && input_w == 64 && output_h == 3 && output_w == 5) {
-            kernel_h = 22;
-            kernel_w = 14;
-            stride_h = 21;
-            stride_w = 13;
-            pad_top = 0;  // Asymmetric padding for exact PyTorch match
-            pad_bottom = 0;
-            pad_left = 1;
-            pad_right = 1;
-        }
-        // For 16×20 → 5×7: use kernel=4×7, stride=3×2 (ensures correct output shape)
-        else if (input_h == 16 && input_w == 20 && output_h == 5 && output_w == 7) {
-            kernel_h = 4;
-            kernel_w = 7;
-            stride_h = 3;
-            stride_w = 2;
-            pad_top = 0;
-            pad_bottom = 0;
-            pad_left = 0;
-            pad_right = 0;
-        }
-        // General pattern: Try the most effective improvements
-        else {
-            // First verify if the base case produces correct output dimensions
-            uint32_t base_output_h = (input_h - kernel_h) / stride_h + 1;
-            uint32_t base_output_w = (input_w - kernel_w) / stride_w + 1;
-            bool base_correct = (base_output_h == output_h) && (base_output_w == output_w);
-
-            // Pattern 1: kernel_h+1 with stride adjustment (often most effective)
-            if ((stride_mismatch || h_non_integer) && !base_correct) {
-                uint32_t test_kernel_h = kernel_h + 1;
-                // Try stride adjustment
-                for (uint32_t pad_h = 0; pad_h <= 4; pad_h++) {
-                    uint32_t padded_h = input_h + pad_h;
-                    if (padded_h >= test_kernel_h) {
-                        uint32_t test_stride_h = (padded_h - test_kernel_h) / (output_h - 1);
-                        if (test_stride_h > 0) {
-                            uint32_t output_h_calc = (padded_h - test_kernel_h) / test_stride_h + 1;
-                            // Also verify width constraint is maintained with existing kernel_w/stride_w
-                            uint32_t output_w_calc = (input_w - kernel_w) / stride_w + 1;
-
-                            if (output_h_calc == output_h && output_w_calc == output_w) {
-                                kernel_h = test_kernel_h;
-                                stride_h = test_stride_h;
-                                if (pad_h > 0) {
-                                    pad_top = pad_h / 2;
-                                    pad_bottom = pad_h - pad_top;
-                                }
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            // Pattern 2: kernel_w+1 when stride mismatch favors width OR when width constraint fails
-            if (!base_correct && (stride_mismatch || w_non_integer || base_output_w != output_w)) {
-                uint32_t test_kernel_w = kernel_w + 1;
-                uint32_t required_input_w = (output_w - 1) * stride_w + test_kernel_w;
-                if (required_input_w <= input_w + 4) {
-                    uint32_t actual_pad_w = (required_input_w > input_w) ? (required_input_w - input_w) : 0;
-                    uint32_t padded_w = input_w + actual_pad_w;
-
-                    // Verify this produces exactly the required output width
-                    uint32_t output_w_calc = (padded_w - test_kernel_w) / stride_w + 1;
-                    if (output_w_calc == output_w) {
-                        kernel_w = test_kernel_w;
-                        if (actual_pad_w > 0) {
-                            pad_left = actual_pad_w / 2;
-                            pad_right = actual_pad_w - pad_left;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    uint32_t actual_padded_h = input_h + pad_top + pad_bottom;
-    uint32_t actual_padded_w = input_w + pad_left + pad_right;
-
-    double memory_overhead =
-        100.0 * (actual_padded_h * actual_padded_w - input_h * input_w) / (double)(input_h * input_w);
-
-    return AdaptivePoolingParams{
-        .kernel_size = {kernel_h, kernel_w},
-        .stride = {stride_h, stride_w},
-        .padding = {pad_top, pad_bottom, pad_left, pad_right},
-        .padded_h = actual_padded_h,
-        .padded_w = actual_padded_w,
-        .memory_overhead_percent = memory_overhead};
-}
-
 // Generic invoke function for both max and avg pool operations. Most of the arguments are shared excpet for the
 // dilation which is set to (1,1) for avg pool and count_include_pad and divisor_override which have no effect on
 // maxpool.
@@ -956,196 +341,6 @@ Tensor AvgPool2DOp::invoke(
         in_place_halo,
         deallocate_input,
         reallocate_halo_output);
-}
-
-Tensor AdaptiveAvgPool2DOp::invoke(
-    QueueId queue_id,
-    const Tensor& input_tensor,
-    uint32_t batch_size,
-    uint32_t input_h,
-    uint32_t input_w,
-    uint32_t channels,
-    std::array<uint32_t, 2> output_size,
-    const std::optional<const MemoryConfig>& memory_config,
-    const std::optional<const TensorMemoryLayout> applied_shard_scheme,
-    bool in_place_halo) {
-    log_debug(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] ENTRY: input_tensor.dtype={}, input_tensor.layout={}",
-        static_cast<int>(input_tensor.dtype()),
-        static_cast<int>(input_tensor.layout()));
-    log_debug(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] PARAMS: batch={}, input_h={}, input_w={}, channels={}, output_size=[{}, {}]",
-        batch_size,
-        input_h,
-        input_w,
-        channels,
-        output_size[0],
-        output_size[1]);
-
-    uint32_t output_h = output_size[0];
-    uint32_t output_w = output_size[1];
-
-    // PATTERN-BASED HYBRID: Analyze kernel patterns and apply padding/dilation as needed
-    log_info(tt::LogOp, "[AdaptiveAvgPool2D] Using PATTERN-BASED HYBRID approach - prioritizing correctness");
-
-    auto hybrid_config = calculate_pattern_based_hybrid_config(input_h, input_w, output_h, output_w);
-    auto params = convert_hybrid_to_legacy(hybrid_config, input_h, input_w);
-
-    // PATTERN-BASED HYBRID RESULTS
-    log_info(tt::LogOp, "[AdaptiveAvgPool2D] === PATTERN-BASED HYBRID RESULTS ===");
-    log_info(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] Strategy: {} | Coverage improvement: {:.1f}%",
-        static_cast<int>(hybrid_config.strategy),
-        hybrid_config.coverage_improvement_percent);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] Variance: {}×{} -> {}×{}",
-        hybrid_config.h_variance,
-        hybrid_config.w_variance,
-        hybrid_config.h_variance_after,
-        hybrid_config.w_variance_after);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] Kernel: {}×{} | Stride: {}×{}",
-        params.kernel_size[0],
-        params.kernel_size[1],
-        params.stride[0],
-        params.stride[1]);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveAvgPool2D] Padding: [{},{},{},{}] | Dilation: {}×{}",
-        params.padding[0],
-        params.padding[1],
-        params.padding[2],
-        params.padding[3],
-        hybrid_config.dilation[0],
-        hybrid_config.dilation[1]);
-    if (params.memory_overhead_percent > 0.0) {
-        log_info(
-            tt::LogOp,
-            "[AdaptiveAvgPool2D] Memory overhead: {:.1f}% | Beneficial: {}",
-            params.memory_overhead_percent,
-            hybrid_config.is_beneficial() ? "YES" : "NO");
-    }
-
-    log_debug(tt::LogOp, "[AdaptiveAvgPool2D] CALLING pool2d_invoke with PATTERN-BASED HYBRID parameters");
-
-    return pool2d_invoke(
-        queue_id,
-        input_tensor,
-        Pool2DType::AVG_POOL2D,
-        batch_size,
-        input_h,
-        input_w,
-        channels,
-        params.kernel_size,
-        params.stride,
-        params.padding,
-        hybrid_config.dilation,  // PATTERN-BASED: Calculated dilation based on kernel analysis
-        false,                   // ceil_mode = false
-        false,                   // count_include_pad = FALSE (ignores padding values)
-        std::nullopt,            // divisor_override
-        memory_config,
-        applied_shard_scheme,
-        in_place_halo);
-}
-
-Tensor AdaptiveMaxPool2DOp::invoke(
-    QueueId queue_id,
-    const Tensor& input_tensor,
-    uint32_t batch_size,
-    uint32_t input_h,
-    uint32_t input_w,
-    uint32_t channels,
-    std::array<uint32_t, 2> output_size,
-    const std::optional<const MemoryConfig>& memory_config,
-    const std::optional<const TensorMemoryLayout> applied_shard_scheme,
-    bool in_place_halo) {
-    log_debug(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] ENTRY: input_tensor.dtype={}, input_tensor.layout={}",
-        static_cast<int>(input_tensor.dtype()),
-        static_cast<int>(input_tensor.layout()));
-    log_debug(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] PARAMS: batch={}, input_h={}, input_w={}, channels={}, output_size=[{}, {}]",
-        batch_size,
-        input_h,
-        input_w,
-        channels,
-        output_size[0],
-        output_size[1]);
-
-    uint32_t output_h = output_size[0];
-    uint32_t output_w = output_size[1];
-
-    // PATTERN-BASED HYBRID: Same approach as AvgPool
-    log_info(tt::LogOp, "[AdaptiveMaxPool2D] Using PATTERN-BASED HYBRID approach - prioritizing correctness");
-
-    auto hybrid_config = calculate_pattern_based_hybrid_config(input_h, input_w, output_h, output_w);
-    auto params = convert_hybrid_to_legacy(hybrid_config, input_h, input_w);
-
-    // PATTERN-BASED HYBRID RESULTS
-    log_info(tt::LogOp, "[AdaptiveMaxPool2D] === PATTERN-BASED HYBRID RESULTS ===");
-    log_info(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] Strategy: {} | Coverage improvement: {:.1f}%",
-        static_cast<int>(hybrid_config.strategy),
-        hybrid_config.coverage_improvement_percent);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] Variance: {}×{} -> {}×{}",
-        hybrid_config.h_variance,
-        hybrid_config.w_variance,
-        hybrid_config.h_variance_after,
-        hybrid_config.w_variance_after);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] Kernel: {}×{} | Stride: {}×{}",
-        params.kernel_size[0],
-        params.kernel_size[1],
-        params.stride[0],
-        params.stride[1]);
-    log_info(
-        tt::LogOp,
-        "[AdaptiveMaxPool2D] Padding: [{},{},{},{}] | Dilation: {}×{}",
-        params.padding[0],
-        params.padding[1],
-        params.padding[2],
-        params.padding[3],
-        hybrid_config.dilation[0],
-        hybrid_config.dilation[1]);
-    if (params.memory_overhead_percent > 0.0) {
-        log_info(
-            tt::LogOp,
-            "[AdaptiveMaxPool2D] Memory overhead: {:.1f}% | Beneficial: {}",
-            params.memory_overhead_percent,
-            hybrid_config.is_beneficial() ? "YES" : "NO");
-    }
-
-    log_debug(tt::LogOp, "[AdaptiveMaxPool2D] CALLING pool2d_invoke with PATTERN-BASED HYBRID parameters");
-
-    return pool2d_invoke(
-        queue_id,
-        input_tensor,
-        Pool2DType::MAX_POOL2D,
-        batch_size,
-        input_h,
-        input_w,
-        channels,
-        params.kernel_size,
-        params.stride,
-        params.padding,
-        hybrid_config.dilation,  // PATTERN-BASED: Calculated dilation based on kernel analysis for MaxPool too
-        false,                   // ceil_mode = false
-        false,                   // count_include_pad = FALSE (consistent with avg pool approach)
-        std::nullopt,            // divisor_override
-        memory_config,
-        applied_shard_scheme,
-        in_place_halo);
 }
 
 }  // namespace operations::pool
