+++ ttnn/cpp/ttnn/operations/pool/generic/generic_pools.hpp
@@ -54,41 +54,9 @@ struct AvgPool2DOp {
         bool reallocate_halo_output = true);
 };
 
-struct AdaptiveAvgPool2DOp {
-    static Tensor invoke(
-        QueueId queue_id,
-        const Tensor& input_tensor,
-        uint32_t batch_size,
-        uint32_t input_h,
-        uint32_t input_w,
-        uint32_t channels,
-        std::array<uint32_t, 2> output_size,
-        const std::optional<const MemoryConfig>& memory_config = std::nullopt,
-        std::optional<const TensorMemoryLayout> applied_shard_scheme = std::nullopt,
-        bool in_place_halo = false);
-};
-
-struct AdaptiveMaxPool2DOp {
-    static Tensor invoke(
-        QueueId queue_id,
-        const Tensor& input_tensor,
-        uint32_t batch_size,
-        uint32_t input_h,
-        uint32_t input_w,
-        uint32_t channels,
-        std::array<uint32_t, 2> output_size,
-        const std::optional<const MemoryConfig>& memory_config = std::nullopt,
-        std::optional<const TensorMemoryLayout> applied_shard_scheme = std::nullopt,
-        bool in_place_halo = false);
-};
-
 }  // namespace operations::pool
 
 constexpr auto max_pool2d = ttnn::register_operation<"ttnn::max_pool2d", operations::pool::MaxPool2DOp>();
 constexpr auto avg_pool2d = ttnn::register_operation<"ttnn::avg_pool2d", operations::pool::AvgPool2DOp>();
-constexpr auto adaptive_avg_pool2d =
-    ttnn::register_operation<"ttnn::adaptive_avg_pool2d", operations::pool::AdaptiveAvgPool2DOp>();
-constexpr auto adaptive_max_pool2d =
-    ttnn::register_operation<"ttnn::adaptive_max_pool2d", operations::pool::AdaptiveMaxPool2DOp>();
 
 }  // namespace ttnn
